+++ backend/src/main/java/com/example/carlos/entities/Client.java
@@ -25,7 +25,7 @@ public class Client implements Serializable {
 	private String cpf;
 	private String email;
 	private String address;
-	private Long phoneNumber;
+	private Integer phoneNumber;
 	private String gender;
 
 	@Column(columnDefinition = "TIMESTAMP WITHOUT TIME ZONE")
@@ -40,7 +40,7 @@ public class Client implements Serializable {
 	public Client() {
 	}
 
-	public Client(Long id, String name, String cpf, String email, String address, Long phoneNumber, String gender,
+	public Client(Long id, String name, String cpf, String email, String address, Integer phoneNumber, String gender,
 			Instant birthDate) {
 		this.id = id;
 		this.name = name;
@@ -92,11 +92,11 @@ public class Client implements Serializable {
 		this.address = address;
 	}
 
-	public Long getPhoneNumber() {
+	public Integer getPhoneNumber() {
 		return phoneNumber;
 	}
 
-	public void setPhoneNumber(Long phoneNumber) {
+	public void setPhoneNumber(Integer phoneNumber) {
 		this.phoneNumber = phoneNumber;
 	}
